requirements.txt
anyio==3.6.1
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
fastapi==0.79.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
mongoengine==0.24.2
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pymongo==4.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
Werkzeug==2.1.2
zipp==3.7.0
